+++ src/components/domains/shared/index.ts
@@ -1,0 +1,3 @@
+export { default as DefaultFooter } from './DefaultFooter';
+export { default as DefaultHero } from './DefaultHero';
+export { default as DefaultNav } from './DefaultNav';