+++ src/components/layout/index.tsx
@@ -1,3 +1,5 @@
 export { Container, type ContainerProps } from './Container';
+export { default as DomainFooter } from './DomainFooter';
+export { default as DomainNavigation } from './DomainNavigation';
 export { Footer } from './Footer';
 export { Navigation } from './Navigation';